requirements.txt
strands-agents[gemini]>=1.0.0
strands-agents-tools>=0.2.0
